enable host only permissions
